Fix for size calculation.
git-svn-id: http://ebookdroid.googlecode.com/svn/trunk@187 714301b5-5bf2-3aa6-4ccc-ecebab0412e3

+++ src/org/ebookdroid/core/DecodeServiceBase.java
@@ -145,12 +145,12 @@ public class DecodeServiceBase implements DecodeService {
         return Math.round(getScaledWidth(vuPage, scale) * currentDecodeTask.pageSliceBounds.width());
     }
 
-    private int getScaledHeight(final CodecPage vuPage, final float scale) {
-        return (int) (scale * vuPage.getHeight());
+    private float getScaledHeight(final CodecPage vuPage, final float scale) {
+        return (scale * vuPage.getHeight());
     }
 
-    private int getScaledWidth(final CodecPage vuPage, final float scale) {
-        return (int) (scale * vuPage.getWidth());
+    private float getScaledWidth(final CodecPage vuPage, final float scale) {
+        return (scale * vuPage.getWidth());
     }
 
     private float calculateScale(final CodecPage codecPage, final int targetWidth) {
